Added unit feedingCountdown and death on 0 health, minor bug fixes

+++ controllers/units.js
@@ -10,7 +10,7 @@ const {
   healthLost,
   unfeedableInterval,
   timeout,
-} = require('./config_intervals');
+} = require('./config_values');
 
 const getAllUnits = async (req, res) => {
   res.send('get all units');
@@ -25,7 +25,7 @@ const getUnit = asyncWrapper(async (req, res, next) => {
   res.status(StatusCodes.OK).json({ unit });
 });
 
-const addUnitToStable = asyncWrapper(async (req, res, next) => {
+const addUnitToBuilding = asyncWrapper(async (req, res, next) => {
   const { type, buildingId } = req.body;
   const building = await Building.findByPk(buildingId);
   if (!building) {
@@ -45,6 +45,31 @@ const addUnitToStable = asyncWrapper(async (req, res, next) => {
     feedable: true,
   };
   const result = await Unit.create(unit);
+
+  const feedingCountdown = setInterval(async () => {
+    const unitToUpdate = await Unit.findByPk(result.id);
+    if (unitToUpdate.health - healthLost <= 0) {
+      await Unit
+        .update(
+          { health: 0, alive: false, feedable: false },
+          { where: { id: unitToUpdate.id } },
+        );
+      // eslint-disable-next-line max-len
+      const buildingToUpdate = await Building.findByPk(buildingId); // Called in the case of the number of units in a building changing during the interval so we have the updated value
+      await Building
+        .update(
+          { numberOfUnits: buildingToUpdate.numberOfUnits - 1 },
+          { where: { id: buildingToUpdate.id } },
+        );
+      clearInterval(feedingCountdown);
+    } else {
+      await Unit
+        .update(
+          { health: unitToUpdate.health - healthLost },
+          { where: { id: unitToUpdate.id } },
+        );
+    }
+  }, unitFeedingInterval);
   const resultBuilding = await Building
     .update({ numberOfUnits: building.numberOfUnits + 1 }, { where: { id: buildingId } });
   res.status(StatusCodes.CREATED).json({ result, resultBuilding });
@@ -64,11 +89,23 @@ const feedUnit = asyncWrapper(async (req, res, next) => { // feed unit
   res.status(StatusCodes.OK).json({ result, result2, success: true });
 });
 
-const deleteUnit = asyncWrapper(async (req, res, next) => {
+const deleteUnit = asyncWrapper(async (req, res, next) => { // ***NEEDS TO CLEAR INTERVAL***
   const { id: unitID } = req.params;
+
+  const unit = await Unit.findByPk(unitID);
+  const buildingToUpdate = await Building.findByPk(unit.BuildingId);
+  if (!buildingToUpdate) {
+    return next(createCustomError(`No building with id : ${unit.BuildingId} where unit with id: ${unitID} is currently located`, StatusCodes.NOT_FOUND));
+  }
+  await Building
+    .update(
+      { numberOfUnits: buildingToUpdate.numberOfUnits - 1 },
+      { where: { id: buildingToUpdate.id } },
+    );
+
   const result = await Unit.destroy({ where: { id: unitID } });
   if (!result) {
-    return next(createCustomError(`No unit with id : ${unitID}`, 404));
+    return next(createCustomError(`No unit with id : ${unitID}`, StatusCodes.NOT_FOUND));
   }
   res.status(200).json({ result });
 });
@@ -76,7 +113,7 @@ const deleteUnit = asyncWrapper(async (req, res, next) => {
 module.exports = {
   getAllUnits,
   getUnit,
-  addUnitToStable,
+  addUnitToBuilding,
   feedUnit,
   deleteUnit,
 };
